move finfo mime tests to UtilMimeTests.php
added fallback test method added
