+++ tests/UtilTests.php
@@ -98,22 +98,4 @@ class UtilTests extends \PHPUnit_Framework_TestCase
         $result = Util::normalizePath($input);
         $this->assertEquals($expected, $result);
     }
-
-    public function pathAndContentProvider()
-    {
-        return array(
-            array('/some/file.css', 'body { background: #000; } ', 'text/css'),
-            array('/some/file.txt', 'body { background: #000; } ', 'text/plain'),
-            array('/1x1', base64_decode('R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs='), 'image/gif')
-        );
-    }
-
-    /**
-     * @dataProvider  pathAndContentProvider
-     */
-    public function testGuessMimeType($path, $content, $expected)
-    {
-        $mimeType = Util::guessMimeType($path, $content);
-        $this->assertEquals($expected, $mimeType);
-    }
 }
